HAE - Add support for sending events and key-presses to UInput
UInput is a linux kernel module that allows to handle the input
subsystem from user land.
The goal of this integration in samples, is to allow the injection of
real events in linux kernel.
By default, this integration is not activated.
To activate it, build HAE with 'UINPUT=on' option in SCONS build
command line.

Change-Id: I221ba15f76728977ee49a1954a2ba82b8d0083e0
Signed-off-by: Enric GUIHO <eguiho.ext@orange.com>
Signed-off-by: Fabrice FONTAINE <fabrice.fontaine@orange.com>