ASAHAE-26 remove gradle wrapper
The gradle wrapper is a binary licensed
under ASL2.0 which is more restrictive
than the license for alljoyn etc.

This patch removes the binary and on import/open
of the project in Android Studio, The user is prompted
to setup the wrapper locally.

Change-Id: I74a14e8f22f27fd11fc79086ab5061873aef5237
Signed-off-by: Nathaniel Garbutt <nathaniel@two-bulls.com>

ASAHAE-26 Remove gradle wrapper

added gradle wrapper to .gitignore
removed iml files as they are generated by gradle.
added *.iml to .gitignore

Change-Id: I83ea40e394e98a174359699bac4f80b2190b43cf
Signed-off-by: Nathaniel Garbutt <nathaniel@two-bulls.com>