ASABASE-1069 Update CDM Controller C++ code so that it matches XML definitions
This is a large commit that ensures the C++ controller code has names and types that match the XML definitions.
For example:
- Some C++ interfaces exposed a method called "GetCyclePhasesDescriptions" when the XML file calls the method "GetVendorPhasesDescription".
- Some properties were typed as integers in C++ when the XML file defines them as enums.
- Some types were not consistent. (Example: C++ was WaterLevelSupplySource, XML is SupplySource)

This commit also takes the opportunity to clean up controller files so ordering and comments are consistent which allows our automated verification and refactoring tools to work more effectively. This should provide a solid base for future 16.10 controller work.

Other than type and name changes, there should be no function changes. The behaviour of the code should be exactly the same.

Change-Id: I2f6b28f0f21db9e77cd4cde50e766e8e1d21588f
Signed-off-by: James Lee <james.lee@two-bulls.com>