ASABASE-1067 CreateInterface returns null when interface is already created
CdmController::CreateInterface now returns a shared_ptr, this allows us to remove
a lot of house keeping inside Cdm that tried to do automatic memory managment.
Each Interface now has its own Proxy Bus Object. It's cheap to make PBOs,
but we might want to make a pool of them for some reason,
because of this they're passed into the interface as a shared_ptr

Change-Id: Ia1e9b56f93aafb9712555050ff4ffea4efcb6984
Signed-off-by: Nathaniel Garbutt <nathaniel@two-bulls.com>