ASABASE-1110 Sample launcher: CDM OS
CDM OS is a python program that is used to launch std and tcl cdm samples.
It provides a some 'boot programs', including one that simulates a shell,
that can be used to manipulate the emulated devices state in a local manner.
The sample is then alerted to the state change and emits a signal to controllers.

This patch also contains some code gen bug fixes for tcl
and removes that static data that was stored inside the c++ models. This data
now lives in property files and can be manipulated with CDM OS.

Change-Id: Ib69f18f908a93545c18a04e7856932422c5762b9
Signed-off-by: Nathaniel Garbutt <nathaniel@two-bulls.com>