ASABASE-1090 LSF translator
The LSF translator has been hooked up into the create interface method.
The implementation covers On/Off, Brightness, Hue and Saturation.

There is a generic framework base class that can be used to implement
other translators. The LSF implementation is currently experimental
and is included as private within the framework. More work needs
to be done and eventually the translators could/should live in a separate
path and be included as plugins (the same as interfaces)

Change-Id: I07ca840930be4552ab37bd1080aa51da9c1ec954
Signed-off-by: Nathaniel Garbutt <nathaniel@two-bulls.com>