ASABASE-1117 Cleanup
Fixed Lamp Device sample so it's HAL directories match the interface names
This is nice for CDM OS

Moved the devicexml folder to samples and renamed it.
Device xmls are now installed into the dist dir.

Fixed the device xmls themselves to match the device type.
Consolidated the bus paths and added default property values

Change-Id: Id3a3595e9d04b9219ffbbc3e6886c72755d5bac0
Signed-off-by: Nathaniel Garbutt <nathaniel@two-bulls.com>