ASABASE-1110 CDM OS
I've stripped away a number of arguments to keep it bare bones.
It only used --device --devicexml and --system. All args are optional.

If no device is specified, it will prompt. If the device requires and xml, it will prompt.
system defaults to cdm

Added a monitor to cdm os to monitor the device process.
If the device process exits...then cdm os will abort too.

Change-Id: Id420b1a0d11e495c2e65659d5bc50549e0072326
Signed-off-by: Nathaniel Garbutt <nathaniel@two-bulls.com>